src/requirements.txt
PySide2~=5.15.2
python-dateutil~=2.8.1